Support libgbm on NVIDIA
NVIDIA now provides support for mesa's libgbm on nvidia-drm by
implementing the gbm backend interface in a proprietary blob.
An additional change is necessary in snapd to pull from the host the
proprietary blob implementing the gbm backend interface for nvidia.
Here set GBM_BACKENDS_PATH to the directory in which snapd will mount
the library `nvidia-drm_gbm.so`