amdgpu: Add support for path envar
A new patch in libdrm ( https://gitlab.freedesktop.org/mesa/libdrm/-/merge_requests/273 )
allows to specify a set of paths where to
search for the `amdgpu.ids` file. This patch configures the envar
to search for the file directly in the mesa-2404 snap, and inside
the posible locations inside the snap itself. This patch removes
the need to bind the `amdgpu.ids` file into /usr/share/libdrm.